Divide backtrace in smaller files
commit-id:a1319e9f
